Rename local geometry modules
