+++ pyrokinetics/local_geometry/fourier_cgyro.py
@@ -3,9 +3,9 @@ from typing import Tuple
 from scipy.optimize import least_squares  # type: ignore
 from scipy.integrate import simpson
 from ..constants import pi
-from .LocalGeometry import LocalGeometry
+from .local_geometry import LocalGeometry
 from ..typing import ArrayLike
-from .LocalGeometry import default_inputs
+from .local_geometry import default_inputs
 
 
 def default_fourier_cgyro_inputs():
