+++ pyrokinetics/local_geometry/miller_turnbull.py
@@ -2,9 +2,9 @@ import numpy as np
 from typing import Tuple
 from scipy.optimize import least_squares  # type: ignore
 from ..constants import pi
-from .LocalGeometry import LocalGeometry
+from .local_geometry import LocalGeometry
 from ..typing import ArrayLike
-from .LocalGeometry import default_inputs
+from .local_geometry import default_inputs
 
 
 def default_miller_turnbull_inputs():
